+++ landingzones/views_taskflow.py
@@ -9,6 +9,7 @@ from django.urls import reverse
 from rest_framework.response import Response
 
 # Projectroles dependency
+from projectroles.app_settings import AppSettingAPI
 from projectroles.email import send_generic_mail
 from projectroles.models import Project
 from projectroles.plugins import get_backend_api
@@ -21,16 +22,14 @@ from samplesheets.tasks import update_project_cache_task
 from landingzones.models import LandingZone
 
 
-# Access Django user model
 User = auth.get_user_model()
-
-# Get logger
 logger = logging.getLogger(__name__)
+app_settings = AppSettingAPI()
 
 
 # Local constants
-APP_NAME = 'samplesheets'
-EMAIL_MESSAGE_MOVED = r'''
+APP_NAME = 'landingzones'
+EMAIL_MSG_MOVED = r'''
 Data was successfully validated and moved into the project
 sample data repository from your landing zone.
 
@@ -46,7 +45,7 @@ Zone UUID: {zone_uuid}
 
 Status message:
 "{status_info}"'''.lstrip()
-EMAIL_MESSAGE_FAILED = r'''
+EMAIL_MSG_FAILED = r'''
 Validating and moving data from your landing zone into the
 project sample data repository has failed. Please verify your
 data and request for support if the problem persists.
@@ -62,6 +61,15 @@ Zone UUID: {zone_uuid}
 
 Status message:
 "{status_info}"'''.lstrip()
+EMAIL_MSG_MEMBER = r'''
+{user} ({user_email}) has uploaded {file_count} file{file_count_suffix}
+into assay "{assay}"
+under the project "{project}".
+
+You can browse the assay metadata and related files at
+the following URL:
+{url}
+'''.lstrip()
 
 
 # TODO: Integrate Taskflow API with general SODAR API (see sodar_core#47)
@@ -90,11 +98,8 @@ class TaskflowZoneCreateAPIView(BaseTaskflowAPIView):
 class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
     """API view for setting landing zone status after taskflow operation"""
 
-    def _add_owner_alert(self, zone, file_count, validate_only):
+    def _add_owner_alert(self, app_alerts, zone, file_count, validate_only):
         """Add app alert for zone owner"""
-        app_alerts = get_backend_api('appalerts_backend')
-        if not app_alerts:
-            return
         if zone.status == 'MOVED':
             alert_msg = 'Successfully moved files from landing zone'
             alert_level = 'SUCCESS'
@@ -141,10 +146,33 @@ class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
             project=zone.project,
         )
 
+    def _add_member_alert(self, app_alerts, zone, user, file_count):
+        """Add app alert for project member"""
+        alert_msg = (
+            '{} file{} uploaded by {} under assay "{}" in '
+            'project "{}"'.format(
+                file_count,
+                's' if file_count != 1 else '',
+                zone.user.username,
+                zone.assay.get_display_name(),
+                zone.project.title,
+            )
+        )
+        app_alerts.add_alert(
+            app_name=APP_NAME,
+            alert_name='zone_move_member',
+            user=user,
+            level='INFO',
+            url=reverse(
+                'samplesheets:project_sheets',
+                kwargs={'project': zone.project.sodar_uuid},
+            ),
+            message=alert_msg,
+            project=zone.project,
+        )
+
     def _send_owner_email(self, zone, request):
         """Send email to zone owner"""
-        if not settings.PROJECTROLES_SEND_EMAIL:
-            return
         server_host = settings.SODAR_API_DEFAULT_HOST.geturl()
         subject_body = 'Landing zone {}: {} / {}'.format(
             zone.status.lower(),
@@ -152,7 +180,7 @@ class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
             zone.title,
         )
         if zone.status == 'MOVED':
-            message_body = EMAIL_MESSAGE_MOVED
+            message_body = EMAIL_MSG_MOVED
             email_url = (
                 server_host
                 + reverse(
@@ -163,7 +191,7 @@ class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
                 + str(zone.assay.sodar_uuid)
             )
         else:  # FAILED
-            message_body = EMAIL_MESSAGE_FAILED
+            message_body = EMAIL_MSG_FAILED
             email_url = (
                 server_host
                 + reverse(
@@ -185,7 +213,36 @@ class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
         )
         send_generic_mail(subject_body, message_body, [zone.user], request)
 
+    def _send_member_email(self, zone, request, file_count):
+        """Send member email on landing zone move"""
+        server_host = settings.SODAR_API_DEFAULT_HOST.geturl()
+        subject_body = 'Files uploaded in project "{}" by {}'.format(
+            zone.project.title,
+            zone.user.get_full_name(),
+        )
+        message_body = EMAIL_MSG_MEMBER
+        email_url = (
+            server_host
+            + reverse(
+                'samplesheets:project_sheets',
+                kwargs={'project': zone.project.sodar_uuid},
+            )
+            + '#/assay/'
+            + str(zone.assay.sodar_uuid)
+        )
+        message_body = message_body.format(
+            project=zone.project.title,
+            assay=zone.assay.get_display_name(),
+            user=zone.user.get_full_name(),
+            user_email=zone.user.email,
+            file_count=file_count,
+            file_count_suffix='s' if file_count != 1 else '',
+            url=email_url,
+        )
+        send_generic_mail(subject_body, message_body, [zone.user], request)
+
     def post(self, request):
+        app_alerts = get_backend_api('appalerts_backend')
         try:
             zone = LandingZone.objects.get(sodar_uuid=request.data['zone_uuid'])
         except LandingZone.DoesNotExist:
@@ -204,17 +261,37 @@ class TaskflowZoneStatusSetAPIView(BaseTaskflowAPIView):
         file_count = int(request.data.get('file_count', 0))
         validate_only = bool(int(request.data.get('validate_only', '0')))
 
-        # Create alerts and emails
+        # Create alert and send email for zone owner
         if (
             (zone.status == 'MOVED' and file_count > 0)
             or validate_only
             or zone.status == 'FAILED'
         ):
-            self._add_owner_alert(zone, file_count, validate_only)
-            if not validate_only:
+            if app_alerts:
+                self._add_owner_alert(
+                    app_alerts, zone, file_count, validate_only
+                )
+            if settings.PROJECTROLES_SEND_EMAIL and not validate_only:
                 self._send_owner_email(zone, request)
-            # TODO: If moved, add app alerts to other users
-            # TODO: If moved, send update email to other project users
+
+        # Create alerts and send emails to other project members if enabled
+        member_notify = app_settings.get_app_setting(
+            APP_NAME, 'member_notify_move', project=zone.project
+        )
+        if member_notify and zone.status == 'MOVED' and file_count > 0:
+            members = [
+                r for r in zone.project.get_all_roles() if r.user != zone.user
+            ]
+            for member_role in members:
+                if app_alerts:
+                    self._add_member_alert(
+                        app_alerts=app_alerts,
+                        zone=zone,
+                        user=member_role.user,
+                        file_count=file_count,
+                    )
+                if settings.PROJECTROLES_SEND_EMAIL:
+                    self._send_member_email(zone, request, file_count)
 
         # If zone is removed by moving or deletion, call plugin function
         if request.data['status'] in ['MOVED', 'DELETED']: