+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Added
     - ``busyzones`` management command (#1212)
     - App alerts for sheet cache updates (#1000)
     - App alerts for zone owner for zone actions (#1204)
+    - ``member_notify_move`` app setting (#1203)
+    - Project member notifications from zone moving (#1203)
 - **Samplesheets**
     - Simple link support for string cell rendering (#1001)
     - ``generic_raw`` assay plugin (#1128)